Stationary Sheath

StationarySheath.Feb

StationarySheath.Mesh

StationarySheath.Params

StationarySheath.Phidata_grid

StationarySheath.Phidata_poly

StationarySheath.build_fe!

StationarySheath.build_fi!

StationarySheath.compute_ne!

StationarySheath.compute_ni!

StationarySheath.diags_end

StationarySheath.diags_loop

StationarySheath.eval_fe!

StationarySheath.eval_phi

StationarySheath.eval_phi

StationarySheath.eval_phi_dx

StationarySheath.eval_phi_dx

StationarySheath.evaluate_poly

StationarySheath.get_v_char_e

StationarySheath.get_v_char_i

StationarySheath.integral_char_00!

StationarySheath.integral_char_vb0!

StationarySheath.integral_char_xb0!

StationarySheath.ploot

StationarySheath.solve_Poisson!

StationarySheath.solve_Poisson!

StationarySheath.𝓛ᵢ

StationarySheath.𝓛ₑ

StationarySheath.Feb — Type

mutable struct Feb

Boundary condition $f_{e,b}$ for the electron density at $(x=0,v⩽0)$.

Le_threshold::Float64

Maximal electric energy $\mathcal{L}_e(x,v)$ s.t. $(x,v)\in$supp$(f_{e,b})$

value::Function

Prescribed value at $(x=0,v\leqslant 0)$

StationarySheath.Mesh — Type

struct Mesh

min::Float64

Minimal value

max::Float64

Maximal value

step::Float64

Step

NN::Int64

Number of intervals

mesh::Vector{Float64}

Mesh itself

StationarySheath.Params — Type

struct Params

λ::Float64

Debye length

μ::Float64

Mass ratio

ν::Float64

Ionization factor

meshx::Mesh

Space mesh

Nvi::Int64

Number of points in ion speed mesh

Nve::Int64

Number of points in electron speed mesh

Nit_max::Int64

Maximal number of fixed-point iterations

StationarySheath.Phidata_grid — Type

struct Phidata_grid <: Phidata

Representation of ϕ on a 1D regular mesh.

values::Vector{Float64}

Pointwise values on the space mesh

StationarySheath.Phidata_poly — Type

struct Phidata_poly <: Phidata

Polynomial representation of ϕ. Basis are * canonical: starting from degree 2. * pair: only pair expononents of the canonical basis.

cardinal::Int64

Cardinal of the basis

basis::String

Basis choice

horner::Vector{Vector{Float64}}

Array of Horner coefficients for x -> p(x)

horner_dx::Vector{Vector{Float64}}

Array of Horner coefficients for x -> p'(x)

Vandermonde::Matrix{Float64}

Matrix of the linear system for the Poisson solver

coeffs::Vector

Coefficients in the basis

StationarySheath.build_fe! — Method

build_fe!(fe, params, f_eb, phix)

Evaluation of ``f_e`` on the regular grid meshx × meshve. 

The speed mesh is computed to fit 0 ⩽ 𝓛ₑ ⩽ 𝓛ₑ(1,0).

StationarySheath.build_fi! — Method

build_fi!(fi, params, f_eb, phix, phidx)

Approximation of ``f_i`` on a regular grid meshx × meshvi.

StationarySheath.compute_ne! — Method

compute_ne!(ne, chare, tointe, params, f_eb, phix)

Computation of 
```math
    n_e(x) = \int_{v\in\mathbb{R}} f_e(x,v) dv = 2 \int_{v \leqslant 0} f_e(x,v) dv.
```

StationarySheath.compute_ni! — Method

compute_ni!(
    ni,
    chari,
    vi,
    vip,
    toint,
    params,
    f_eb,
    phix,
    phidx
)

Computation of 
```math
    n_i(x) = \int_{v\in\mathbb{R}} f_i(x,v) dv = 2 \int_{v \leqslant 0} f_i(x_b(x,v),v_b(x,v)) dv.
```

StationarySheath.diags_end — Method

diags_end(params, phi, f_eb, ni, ne, phix, phidx, n)

Aggregation of diagnostics performed at the end of the program.

StationarySheath.diags_loop — Method

diags_loop(params, phi, ni, ne, phix, phixp, n, eps)

Aggregation of diagnostics performed during the fixed-point loop.

StationarySheath.eval_fe! — Method

eval_fe!(res, f_eb, params, phix, v)

Return the value ``f_e(x,v)``. 

Note that the non-emitting boundary condition has precedence over
the prescribed value ``f_{e,b}``, and the result will be 0
whenever 𝓛ₑ(x,v) ⩾ 𝓛ₑ(1,0).

StationarySheath.eval_phi — Method

eval_phi(phi, params, x)

Value of ϕ at `x`.

StationarySheath.eval_phi — Method

eval_phi(phi, params, x)

Value of ϕ at `x`.

StationarySheath.eval_phi_dx — Method

eval_phi_dx(phi, params, x)

Value of ϕ' at `x`.

StationarySheath.eval_phi_dx — Method

eval_phi_dx(phi, params, x)

Value of ϕ' at `x`.

StationarySheath.evaluate_poly — Method

evaluate_poly(eval, x)

Return the polynomial 
```math
    a_1 + x (a_2 + x \dots (a_{n-1} + x a_n))
```
with `eval```=[a_1,\dots,a_n]``.

StationarySheath.get_v_char_e — Method

get_v_char_e(params, v0, phix0, phix)

Return the negative ``v`` such that
```math
    \frac{v^2}{2} - \frac{1}{\mu} \phi(x) = \frac{v_0^2}{2} - \frac{1}{\mu} \phi(x_0)
```

StationarySheath.get_v_char_i — Method

get_v_char_i(v0, phix0, phix)

Return the negative ``v`` such that
```math
    \frac{v^2}{2} + \phi(x) = \frac{v_0^2}{2} + \phi(x_0)
```

StationarySheath.integral_char_00! — Method

integral_char_00!(vi, toint, params, phix, f_eb)

Special case of the critical characteristic:
use of a rectangle integration scheme to avoid the equilibrium point ``(x=0,v=0)``.

StationarySheath.integral_char_vb0! — Method

integral_char_vb0!(
    vi,
    toint,
    params,
    ixb,
    phix,
    phidx,
    f_eb
)

Compute the integral of f_e along the ion characteristic reaching ``(x_b>0,v_b=0)`` at ``\tau=0``.
Here ``x_b`` is known through ``\phi(x_b)`` only. Uses 
```math
    \int_{\tau=-\infty}^{0} f_e(x(\tau),v(\tau))
    = \int_{v=vmin}^{0} \frac{f_e(x(v),v)}{-\phi'(x(v))} dv
```
and a trapeze integral on the points ``(x(v),v)`` s.t. ``x(v)\in``meshx.
Preallocated vectors vi and toint will be modified by the function.

StationarySheath.integral_char_xb0! — Method

integral_char_xb0!(vi, toint, params, vb, phix, f_eb)

Compute the integral of f_e along the ion characteristic reaching ``(x_b=0,v_b<0)`` at ``\tau=0``.
Uses 
```math
    \int_{\tau=-\infty}^{0} f_e(x(\tau),v(\tau))
    = \int_{z=0}^{1} \frac{f_e(z,-\sqrt{v_b^2 - 2*\phi(z)})}{-\sqrt{v_b^2 - 2*\phi(z)}} dz
```
and a trapeze integral on the points ``(z,-g_z(v_b))`` s.t. ``z\in``meshx.
Preallocated vectors vi and toint will be modified by the function.

StationarySheath.ploot — Method

ploot(mesh, value, tag; erase)

Plot and wait for user input.

StationarySheath.solve_Poisson! — Method

solve_Poisson!(phi, params, ni, ne)

Resolution by quadrature when ``\phi`` is approximated pointwise.

StationarySheath.solve_Poisson! — Method

solve_Poisson!(phi, params, ni, ne)

Resolution by least squares when ``\phi`` is decomposed in a polynomial basis.

StationarySheath.𝓛ᵢ — Method

𝓛ᵢ(phix, v)

Ion infinitesimal energy.

StationarySheath.𝓛ₑ — Method

𝓛ₑ(params, phix, v)

Electric infinitesimal energy.